Extract sites display settings to its own component
